UTs for checking if loop mechanisms exits when their controllers are dropped

+++ crates/client/src/protocol_consts.rs
@@ -33,7 +33,7 @@ impl From<i64> for ListOffsetsTimestampType {
             -1 => ListOffsetsTimestampType::Latest,
             o => {
                 assert!(o >= 0, "Unknown ListOffsetsTimestampType value");
-                Specific(o as i64)
+                Specific(o)
             }
         }
     }
